+++ src/hooks/useAuth.ts
@@ -1,0 +1,40 @@
+'use client';
+
+import { useState, useEffect } from 'react';
+import { useRouter } from 'next/navigation';
+
+export function useAuth() {
+  const [isAuthenticated, setIsAuthenticated] = useState(false);
+  const [isLoading, setIsLoading] = useState(true);
+  const router = useRouter();
+
+  useEffect(() => {
+    // 認証状態をチェック
+    const token = localStorage.getItem('token');
+    setIsAuthenticated(!!token);
+    setIsLoading(false);
+  }, []);
+
+  const login = (token: string, user?: any) => {
+    localStorage.setItem('token', token);
+    if (user) {
+      localStorage.setItem('user', JSON.stringify(user));
+    }
+    setIsAuthenticated(true);
+    router.push('/po/upload');
+  };
+
+  const logout = () => {
+    localStorage.removeItem('token');
+    localStorage.removeItem('user');
+    setIsAuthenticated(false);
+    router.push('/po/login');
+  };
+
+  return {
+    isAuthenticated,
+    isLoading,
+    login,
+    logout,
+  };
+}
